Move fetch_projects to utils
