+++ rickchurch/church.py
@@ -1,7 +1,6 @@
 import logging
 from typing import Any, Callable, Dict, List, Optional
 
-import asyncpg
 import fastapi
 import pydispix
 from fastapi.openapi.utils import get_openapi
@@ -10,6 +9,7 @@ from rickchurch import constants
 from rickchurch.auth import authorized
 from rickchurch.log import setup_logging
 from rickchurch.models import Project
+from rickchurch.utils import fetch_projects
 
 logger = logging.getLogger("rickchurch")
 app = fastapi.FastAPI()
@@ -85,25 +85,7 @@ async def setup_data(request: fastapi.Request, callnext: Callable) -> fastapi.Re
     return response
 
 
-async def fetch_projects(db_conn: asyncpg.Connection) -> List[Project]:
-    """Obtain list of active projects in the database"""
-    async with db_conn.transaction():
-        db_projects = await db_conn.fetchrow("SELECT * FROM projects")
-
-    projects = []
-    for db_project in db_projects:
-        project = Project(
-            name=db_project["project_name"],
-            x=db_project["position_x"],
-            y=db_project["position_y"],
-            priority=db_project["project_priority"],
-            image=db_project["base64_image"]
-        )
-        projects.append(project)
-    return projects
-
-
-# region: endpoints
+# region: Member Endpoints
 
 @app.get("/get_projects", tags=["Member endpoint"], response_model=List[Project])
 async def get_projects(request: fastapi.Request) -> List[Project]:
